Fix cli! based on new semantics
Apparently, there were some changes introduced in 0.2.2 of
`org.clojure/tools.logging` that affected some of the behavior
of the cli validation that we were doing in our `utils` namespace.

PuppetDB is patching this here:

https://github.com/puppetlabs/puppetdb/pull/751/files

This commit simply brings that same change over to kitchensink,
since it'll be needed when we submit the PR to move PuppetDB
over to kitchensink.