Utility functions to create SSLContext et al from pems
We are starting to see more and more places in the code of other
projects where we need to read in the three common puppet PEM
files (cert, key, ca cert) and return KeyStores representing
the key/trust store, or SSLContext objects for use in
various SSL initialization code.  This commit simply adds
functions to perform these services.  Downstream projects
that contain similar variants of this code should be updated
to use these.