(maint) Define our own entropy in `open-port-num`.
The current implementation of `open-port-num` relies on the JDK
implementation to select a currently-open port. Anecdotally, some
implementations have low amounts of entropy for this operation. Change
the implementation to uniformly select a random port number in the
traditional ephemeral port range of 49152 through 65535, so we're sure
we're using the maximum available range regardless of JDK
implementation.

The anecdotal evidence for the entropy comes from the frequency of port
collisions seen in the classifier's integration tests. Up through
2016.4.x, we used a custom random port selection function with entropy
equivalent to this one (but which did *not* check to see if the selected
port was open), and we saw closed ports in the ~12 ports bound less than
once every fifteen runs. Starting in 2017.2.x, we switched to using
kitchensink's open-port-num, and we've observed the frequency of
collisions to go up dramatically: now we see them, on average, once
every other run.

This anecdotal evidence is backed up by observing that this new
implementation of `open-port-num` returns over twice as many distinct
port numbers as the previous one when invoked 60k times during the
tests: ~15.9k vs. ~7k. However, this is only testing my JDK
implementation; that spread may vary depending on the particular JDK
used.