(maint) Clean up with-additional-classpath-entries test
The implementation of this test was not ideal, and made worse by Java 9 changes
to classloaders. Instead of querying URLs, just add a resource to a directory
that's not normally in the class path and test that we can resolve it when it's
added and that we can't after it gets cleaned up.