(PE-33707) Add a keyword->str function
In serveral services the clojure `name` function is being used for turning keywords in to strings. The behavior of the `name` function https://clojuredocs.org/clojure.core/name around dropping namespace from the key name is not desired in some cases. In some cases we just want a literal string representation of the keyword without the preceeding colon character. This commit introduces a new function to do this transformation.