(PE-33707) Add key->str function to convert keywords to strings
We typically use the `name` function to convert keywords to strings,
which is usually correct but has a subtle bug: if the keyword contains a
slash, `name` only returns the portion after the slash, as the portion
before that is considered the namespace.

This commit adds a new `key->str` function which accepts both keywords
and strings and returns the properly stringified version (with the
leading : removed for keywords). Accepting both keywords and strings
allows it to be used to normalize map keys parsed from JSON without
worrying where they came from.