(maint) add `delete-recursively` function to files namespace
This adds a function `delete-recursively` that given a path to a file or
directory as a string, will attempt to delete the file, or the contents
of the directory tree. If at any point in the traversal a file or directory
can't be removed, the recursion stops, and the function returns false.

Otherwise the function will return true if all the files and directories
could be deleted.