(maint) add puppet setting duration conversion functions
Puppet has a setting option called Duration, which supports a single
number with an optional time descriptor (`y`, `d`, `m`, `s`).

This adds functions that convert the supported puppet duration values
into seconds and java.time.Duration accordingly.

An additional function to parse numbers and strings into numbers
was introduced as it is common in the consuming applications to do
this.